+++ js/web-components/utils-format.js
@@ -68,6 +68,11 @@ export class Format {
             return dateTime;
         }
 
+        // Make sure a value exists
+        if (dateTime === null || dateTime === '') {
+            return null;
+        }
+
         // Return formatted date/time in the user's local language
         try {
             if (dateTime instanceof Date) {